Place hypothesis tests in test_ripser and remove superseded tests there. Add regression test for #465
